@@ -8,9 +8,11 @@ Definition of RSA Dataset class and TemporalDataset
 
 from __future__ import annotations
 from typing import List, Optional
+from warnings import warn
 from copy import deepcopy
 import numpy as np
 from pandas import DataFrame
+from rsatoolbox.data.ops import merge_datasets
 from rsatoolbox.util.data_utils import get_unique_unsorted
 from rsatoolbox.util.data_utils import get_unique_inverse
 from rsatoolbox.util.descriptor_utils import check_descriptor_length_error
@@ -18,7 +20,6 @@ from rsatoolbox.util.descriptor_utils import subset_descriptor
 from rsatoolbox.util.descriptor_utils import num_index
 from rsatoolbox.util.descriptor_utils import format_descriptor
 from rsatoolbox.util.descriptor_utils import parse_input_descriptor
-from rsatoolbox.util.descriptor_utils import append_obs_descriptors
 from rsatoolbox.util.descriptor_utils import desc_eq
 from rsatoolbox.io.hdf5 import read_dict_hdf5
 from rsatoolbox.io.pkl import read_dict_pkl
@@ -180,7 +181,7 @@ class Dataset(DatasetBase):
 
         """
         desc = self.obs_descriptors[by]
-        order = np.argsort(desc)
+        order = np.argsort(desc, kind='stable')
         self.measurements = self.measurements[order]
         self.obs_descriptors = subset_descriptor(self.obs_descriptors, order)
 
@@ -244,8 +245,8 @@ class Dataset(DatasetBase):
         ds_part = self.split_obs(obs_desc)
         odd_list = ds_part[0::2]
         even_list = ds_part[1::2]
-        odd_split = merge_subsets(odd_list)
-        even_split = merge_subsets(even_list)
+        odd_split = merge_datasets(odd_list)
+        even_split = merge_datasets(even_list)
         return odd_split, even_split
 
     def nested_odd_even_split(self, l1_obs_desc, l2_obs_desc):
@@ -282,8 +283,8 @@ class Dataset(DatasetBase):
             odd_split, even_split = partition.odd_even_split(l2_obs_desc)
             odd_list.append(odd_split)
             even_list.append(even_split)
-        odd_split = merge_subsets(odd_list)
-        even_split = merge_subsets(even_list)
+        odd_split = merge_datasets(odd_list)
+        even_split = merge_datasets(even_list)
         return odd_split, even_split
 
     @staticmethod
@@ -700,17 +701,39 @@ class TemporalDataset(Dataset):
         self.measurements = self.measurements[order]
         self.obs_descriptors = subset_descriptor(self.obs_descriptors, order)
 
-    def convert_to_dataset(self, by):
-        """ converts to Dataset long format.
-            time dimension is absorbed into observation dimension
+    def time_as_channels(self) -> Dataset:
+        """Converts this to a standard Dataset "long format",
+        where timepoints are represented as additional channels.
 
         Args:
-            by(String): the descriptor which indicates the time dimension in
-                the time_descriptor
+            by (str): the descriptor which indicates the time dimension in
+                the time_descriptor.
 
         Returns:
             Dataset
+        """
+        n_obs, n_chans, n_tps = self.measurements.shape
+        old_chn_des = self.channel_descriptors
+        chn_des = {k: np.repeat(v, n_tps) for (k, v) in old_chn_des.items()}
+        for k, v in self.time_descriptors.items():
+            chn_des[k] = np.tile(v, n_chans)
+        return Dataset(
+            measurements=self.measurements.reshape(n_obs, -1),
+            descriptors=deepcopy(self.descriptors),
+            obs_descriptors=deepcopy(self.obs_descriptors),
+            channel_descriptors=chn_des
+        )
 
+    def time_as_observations(self, by='time') -> Dataset:
+        """Converts this to a standard Dataset "long format",
+        where timepoints are represented as additional observations.
+
+        Args:
+            by (str): the descriptor which indicates the time dimension in
+                the time_descriptor.
+
+        Returns:
+            Dataset
         """
         time = get_unique_unsorted(self.time_descriptors[by])
 
@@ -749,6 +772,24 @@ class TemporalDataset(Dataset):
                           obs_descriptors=obs_descriptors,
                           channel_descriptors=channel_descriptors)
         return dataset
+
+    def convert_to_dataset(self, by):
+        """ converts to Dataset long format.
+            time dimension is absorbed into observation dimension
+
+        Deprecated: Use `TemporalDataset.time_as_observations()` instead.
+
+        Args:
+            by(String): the descriptor which indicates the time dimension in
+                the time_descriptor
+
+        Returns:
+            Dataset
+
+        """
+        warn('Deprecated: [TemporalDataset.convert_to_dataset()]. Replace by '
+            '[TemporalDataset.time_as_observations()]', DeprecationWarning)
+        return self.time_as_observations(by)
 
     def to_dict(self):
         """ Generates a dictionary which contains the information to
@@ -833,6 +874,8 @@ def merge_subsets(dataset_list):
     (e.g., as generated by the subset_* methods). Assumes that descriptors,
     channel descriptors and number of channels per observation match.
 
+    Deprecated. Use `rsatoolbox.data.ops.merge_datasets` instead.
+
     Args:
         dataset_list (list):
             List containing rsatoolbox datasets
@@ -841,26 +884,6 @@ def merge_subsets(dataset_list):
         merged_dataset (Dataset):
             rsatoolbox dataset created from all datasets in dataset_list
     """
-    assert isinstance(dataset_list, list), "Provided object is not a list."
-    assert "Dataset" in str(type(dataset_list[0])), \
-        "Provided list does not only contain Dataset objects."
-    baseline_ds = dataset_list[0]
-    descriptors = baseline_ds.descriptors.copy()
-    channel_descriptors = baseline_ds.channel_descriptors.copy()
-    measurements = baseline_ds.measurements.copy()
-    obs_descriptors = baseline_ds.obs_descriptors.copy()
-
-    for ds in dataset_list[1:]:
-        assert "Dataset" in str(type(ds)), \
-            "Provided list does not only contain Dataset objects."
-        assert descriptors == ds.descriptors.copy(), \
-            "Dataset descriptors do not match."
-        measurements = np.append(measurements, ds.measurements, axis=0)
-        obs_descriptors = append_obs_descriptors(obs_descriptors,
-                                                 ds.obs_descriptors.copy())
-
-    merged_dataset = Dataset(measurements,
-                             descriptors=descriptors,
-                             obs_descriptors=obs_descriptors,
-                             channel_descriptors=channel_descriptors)
-    return merged_dataset
+    warn('Deprecated: [rsatoolbox.data.dataset.merge_subsets()]. Replace by '
+         '[rsatoolbox.data.ops.merge_datasets()]', DeprecationWarning)
+    return merge_datasets(dataset_list)
