@@ -9,18 +9,27 @@ from collections.abc import Iterable
 from copy import deepcopy
 from typing import TYPE_CHECKING, Optional, Tuple
 import numpy as np
-from rsatoolbox.rdm.rdms import RDMs
 from rsatoolbox.rdm.rdms import concat
+from rsatoolbox.rdm.calc_unbalanced import calc_rdm_unbalanced
 from rsatoolbox.rdm.combine import from_partials
 from rsatoolbox.data import average_dataset_by
 from rsatoolbox.util.rdm_utils import _extract_triu_
+from rsatoolbox.util.build_rdm import _build_rdms
+
 if TYPE_CHECKING:
     from rsatoolbox.data.base import DatasetBase
     from numpy.typing import NDArray
 
 
-def calc_rdm(dataset, method='euclidean', descriptor=None, noise=None,
-             cv_descriptor=None, prior_lambda=1, prior_weight=0.1):
+def calc_rdm(
+        dataset: DatasetBase,
+        method: str = 'euclidean',
+        descriptor: Optional[str] = None,
+        noise: Optional[NDArray] = None,
+        cv_descriptor: Optional[str] = None,
+        prior_lambda: float = 1,
+        prior_weight: float = 0.1,
+        remove_mean: bool = False):
     """
     calculates an RDM from an input dataset
 
@@ -40,6 +49,9 @@ def calc_rdm(dataset, method='euclidean', descriptor=None, noise=None,
             precision matrix used to calculate the RDM
             used only for Mahalanobis and Crossnobis estimators
             defaults to an identity matrix, i.e. euclidean distance
+        remove_mean (bool):
+            whether the mean of each pattern shall be removed before distance calculation.
+            This has no effect on poisson based and correlation distances.
 
     Returns:
         rsatoolbox.rdm.rdms.RDMs: RDMs object with the one RDM
@@ -74,14 +86,14 @@ def calc_rdm(dataset, method='euclidean', descriptor=None, noise=None,
             rdm = from_partials(rdms, descriptor=descriptor)
     else:
         if method == 'euclidean':
-            rdm = calc_rdm_euclidean(dataset, descriptor)
+            rdm = calc_rdm_euclidean(dataset, descriptor, remove_mean)
         elif method == 'correlation':
             rdm = calc_rdm_correlation(dataset, descriptor)
         elif method == 'mahalanobis':
-            rdm = calc_rdm_mahalanobis(dataset, descriptor, noise)
+            rdm = calc_rdm_mahalanobis(dataset, descriptor, noise, remove_mean)
         elif method == 'crossnobis':
             rdm = calc_rdm_crossnobis(dataset, descriptor, noise,
-                                      cv_descriptor)
+                                      cv_descriptor, remove_mean)
         elif method == 'poisson':
             rdm = calc_rdm_poisson(dataset, descriptor,
                                    prior_lambda=prior_lambda,
@@ -101,7 +113,7 @@ def calc_rdm(dataset, method='euclidean', descriptor=None, noise=None,
 def calc_rdm_movie(
         dataset, method='euclidean', descriptor=None, noise=None,
         cv_descriptor=None, prior_lambda=1, prior_weight=0.1,
-        time_descriptor='time', bins=None):
+        time_descriptor='time', bins=None, unbalanced=False):
     """
     calculates an RDM movie from an input TemporalDataset
 
@@ -121,6 +133,8 @@ def calc_rdm_movie(
             dimension in dataset.time_descriptors. Defaults to 'time'.
         bins (array-like): list of bins, with bins[i] containing the vector
             of time-points for the i-th bin. Defaults to no binning.
+        unbalanced (bool): if set to True use calc_rdm_unbalanced,
+            else and by default use calc_rdm
 
     Returns:
         rsatoolbox.rdm.rdms.RDMs: RDMs object with RDM movie
@@ -156,18 +170,31 @@ def calc_rdm_movie(
         rdms = []
         for dat in splited_data:
             dat_single = dat.convert_to_dataset(time_descriptor)
-            rdms.append(calc_rdm(dat_single, method=method,
-                                 descriptor=descriptor, noise=noise,
-                                 cv_descriptor=cv_descriptor,
-                                 prior_lambda=prior_lambda,
-                                 prior_weight=prior_weight))
+            if unbalanced:
+                rdms.append(calc_rdm_unbalanced(
+                    dat_single, method=method,
+                    descriptor=descriptor, noise=noise,
+                    cv_descriptor=cv_descriptor,
+                    prior_lambda=prior_lambda,
+                    prior_weight=prior_weight))
+            else:
+                rdms.append(calc_rdm(
+                    dat_single, method=method,
+                    descriptor=descriptor, noise=noise,
+                    cv_descriptor=cv_descriptor,
+                    prior_lambda=prior_lambda,
+                    prior_weight=prior_weight))
 
         rdm = concat(rdms)
         rdm.rdm_descriptors[time_descriptor] = time
+    rdm.dissimilarity_measure = method
     return rdm
 
 
-def calc_rdm_euclidean(dataset, descriptor=None):
+def calc_rdm_euclidean(
+        dataset: DatasetBase,
+        descriptor: Optional[str] = None,
+        remove_mean: bool = False):
     """
     Args:
         dataset (rsatoolbox.data.DatasetBase):
@@ -175,11 +202,13 @@ def calc_rdm_euclidean(dataset, descriptor=None):
         descriptor (String):
             obs_descriptor used to define the rows/columns of the RDM
             defaults to one row/column per row in the dataset
+        remove_mean (bool):
+            whether the mean of each pattern shall be removed
+            before calculating distances.
     Returns:
         rsatoolbox.rdm.rdms.RDMs: RDMs object with the one RDM
     """
-
-    measurements, desc = _parse_input(dataset, descriptor)
+    measurements, desc = _parse_input(dataset, descriptor, remove_mean)
     sum_sq_measurements = np.sum(measurements**2, axis=1, keepdims=True)
     rdm = sum_sq_measurements + sum_sq_measurements.T \
         - 2 * np.dot(measurements, measurements.T)
@@ -204,14 +233,13 @@ def calc_rdm_correlation(dataset, descriptor=None):
         rsatoolbox.rdm.rdms.RDMs: RDMs object with the one RDM
 
     """
-    ma, desc = _parse_input(dataset, descriptor)
-    ma = ma - ma.mean(axis=1, keepdims=True)
+    ma, desc = _parse_input(dataset, descriptor, remove_mean=True)
     ma /= np.sqrt(np.einsum('ij,ij->i', ma, ma))[:, None]
     rdm = 1 - np.einsum('ik,jk', ma, ma)
     return _build_rdms(rdm, dataset, 'correlation', descriptor, desc)
 
 
-def calc_rdm_mahalanobis(dataset, descriptor=None, noise=None):
+def calc_rdm_mahalanobis(dataset, descriptor=None, noise=None, remove_mean: bool = False):
     """
     calculates an RDM from an input dataset using mahalanobis distance
     If multiple instances of the same condition are found in the dataset
@@ -227,14 +255,17 @@ def calc_rdm_mahalanobis(dataset, descriptor=None, noise=None):
             dataset.n_channel x dataset.n_channel
             precision matrix used to calculate the RDM
             default: identity matrix, i.e. euclidean distance
+        remove_mean (bool):
+            whether the mean of each pattern shall be removed
+            before calculating distances.
 
     Returns:
         rsatoolbox.rdm.rdms.RDMs: RDMs object with the one RDM
 
     """
     if noise is None:
-        return calc_rdm_euclidean(dataset, descriptor)
-    measurements, desc = _parse_input(dataset, descriptor)
+        return calc_rdm_euclidean(dataset, descriptor, remove_mean)
+    measurements, desc = _parse_input(dataset, descriptor, remove_mean)
     noise = _check_noise(noise, dataset.n_channel)
     kernel = measurements @ noise @ measurements.T
     rdm = np.expand_dims(np.diag(kernel), 0) + \
@@ -251,7 +282,7 @@ def calc_rdm_mahalanobis(dataset, descriptor=None, noise=None):
 
 
 def calc_rdm_crossnobis(dataset, descriptor, noise=None,
-                        cv_descriptor=None):
+                        cv_descriptor=None, remove_mean: bool = False):
     """
     calculates an RDM from an input dataset using Cross-nobis distance
     This performs leave one out crossvalidation over the cv_descriptor.
@@ -284,6 +315,9 @@ def calc_rdm_crossnobis(dataset, descriptor, noise=None,
             default: identity matrix, i.e. euclidean distance
         cv_descriptor (String):
             obs_descriptor which determines the cross-validation folds
+        remove_mean (bool):
+            whether the mean of each pattern shall be removed
+            before calculating distances.
 
     Returns:
         rsatoolbox.rdm.rdms.RDMs: RDMs object with the one RDM
@@ -314,6 +348,9 @@ def calc_rdm_crossnobis(dataset, descriptor, noise=None,
                 average_dataset_by(data_train, descriptor)
             measurements_test, _, _ = \
                 average_dataset_by(data_test, descriptor)
+            if remove_mean:
+                measurements_train -= measurements_train.mean(axis=1, keepdims=True)
+                measurements_test -= measurements_test.mean(axis=1, keepdims=True)
             rdm = _calc_rdm_crossnobis_single(
                 measurements_train, measurements_test, noise)
             rdms.append(rdm)
@@ -322,7 +359,10 @@ def calc_rdm_crossnobis(dataset, descriptor, noise=None,
         variances = []
         for i, i_fold in enumerate(cv_folds):
             data = datasetCopy.subset_obs(cv_descriptor, i_fold)
-            measurements.append(average_dataset_by(data, descriptor)[0])
+            ma = average_dataset_by(data, descriptor)[0]
+            if remove_mean:
+                ma -= ma.mean(axis=1, keepdims=True)
+            measurements.append(ma)
             variances.append(np.linalg.inv(noise[i]))
         for i_fold in range(len(cv_folds)):
             for j_fold in range(i_fold + 1, len(cv_folds)):
@@ -452,13 +492,16 @@ def _gen_default_cv_descriptor(dataset, descriptor) -> np.ndarray:
 
 def _parse_input(
             dataset: DatasetBase,
-            descriptor: Optional[str]
+            descriptor: Optional[str],
+            remove_mean: bool = False
         ) -> Tuple[np.ndarray, Optional[np.ndarray]]:
     if descriptor is None:
         measurements = dataset.measurements
         desc = None
     else:
         measurements, desc, _ = average_dataset_by(dataset, descriptor)
+    if remove_mean:
+        measurements = measurements - measurements.mean(axis=1, keepdims=True)
     return measurements, desc
 
 
@@ -487,54 +530,3 @@ def _check_noise(noise, n_channel):
     else:
         raise ValueError('noise(s) must have shape n_channel x n_channel')
     return noise
-
-
-def _build_rdms(
-            utv: NDArray,
-            ds: DatasetBase,
-            method: str,
-            obs_desc_name: str | None,
-            obs_desc_vals: Optional[NDArray] = None,
-            cv: Optional[NDArray] = None,
-            noise: Optional[NDArray] = None
-        ) -> RDMs:
-    rdms = RDMs(
-        dissimilarities=np.array([utv]),
-        dissimilarity_measure=method,
-        rdm_descriptors=deepcopy(ds.descriptors)
-    )
-    if (obs_desc_vals is None) and (obs_desc_name is not None):
-        # obtain the unique values in the target obs descriptor
-        _, obs_desc_vals, _ = average_dataset_by(ds, obs_desc_name)
-
-    if _averaging_occurred(ds, obs_desc_name, obs_desc_vals):
-        orig_obs_desc_vals = np.asarray(ds.obs_descriptors[obs_desc_name])
-        for dname, dvals in ds.obs_descriptors.items():
-            dvals = np.asarray(dvals)
-            avg_dvals = np.full_like(obs_desc_vals, np.nan, dtype=dvals.dtype)
-            for i, v in enumerate(obs_desc_vals):
-                subset = dvals[orig_obs_desc_vals == v]
-                if len(set(subset)) > 1:
-                    break
-                avg_dvals[i] = subset[0]
-            else:
-                rdms.pattern_descriptors[dname] = avg_dvals
-    else:
-        rdms.pattern_descriptors = deepcopy(ds.obs_descriptors)
-    # Additional rdm_descriptors
-    if noise is not None:
-        rdms.descriptors['noise'] = noise
-    if cv is not None:
-        rdms.descriptors['cv_descriptor'] = cv
-    return rdms
-
-
-def _averaging_occurred(
-            ds: DatasetBase,
-            obs_desc_name: str | None,
-            obs_desc_vals: NDArray | None
-        ) -> bool:
-    if obs_desc_name is None:
-        return False
-    orig_obs_desc_vals = ds.obs_descriptors[obs_desc_name]
-    return len(obs_desc_vals) != len(orig_obs_desc_vals)
