@@ -1,4 +1,4 @@
-from .rdm_plot import show_rdm, show_rdm_panel, add_descriptor_x_labels, add_descriptor_y_labels
+from .rdm_plot import show_rdm, show_rdm_panel
 from .scatter_plot import show_scatter, show_2d, show_MDS, show_tSNE, show_iso
 from .model_plot import plot_model_comparison
 from .model_map import map_model_comparison
